requirements.txt
# 核心依賴
numpy>=1.24.0
pandas>=2.0.0
polars>=0.20.0
openai>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
jinja2>=3.1.2
python-dotenv>=1.0.0

# 搜尋和處理
scikit-learn>=1.3.0
python-docx>=1.1.0

# 可視化和分析
matplotlib>=3.7.0
seaborn>=0.12.0

# API 服務（選用）
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Google Cloud AI Platform (可選)
google-cloud-aiplatform>=1.36.0

# 開發工具（選用）
# pytest>=7.0.0
# black>=23.0.0
psutil
textual>=0.40.0,<1.0.0